+++ src/components/constructor/blank-constructor-element/blank-constructor-element.jsx
